Update `RomoDatepicker` component to not use jquery
This updates the `RomoDatepicker` component to not use jquery.
This is part of the effort to no longer require jquery to use
Romo. This removes all of the jquery usage within the
`RomoDatepicker` component except for how it initializes an
indicator text input for its elem. This will be updated when the
indicator text input is updated to no longer use jquery. None of
the other components initialize a datepicker component.

This also updates the event names to be prefixed with
`romoDatepicker` instead of just `datepicker`. This is part of
having everything properly namespaced in romo and ensuring it
should work without issue with other javascript libraries.