Update `RomoDropdown` component to not use jquery
This updates the `RomoDropdown` component to not use jquery. This
is part of the effort to no longer require jquery to use Romo.
This removes all of the jquery usage within the `RomoDropdown`
component and also updates the datepicker, dropdown form, option
list dropdown, picker, select, and select dropdown components to
not use the jquery method for initializing a romo dropdown
component.

This removes using jquery's `unbind` to remove all event
handlers from elems. The `Romo.off` doesn't support unbinding all
event handlers but we also decided that Romo doesn't need to be
this aggressive with its components elems. This allows users to
bind event handlers to the same events that romo dropdown does
(though it's not recommended).

This also updates the event names to be prefixed with
`romoDropdown` instead of just `dropdown` and renames variables as
well. This is part of having everything properly namespaced in
romo and ensuring it should work without issue with other
javascript libraries.