Add `Romo.assign` util, use to merge javascript objects
This adds a `Romo.assign` util that can be used to merge
javascript objects. This is similar to jquery's `extend` only
it uses the `Object.assign` function if available or it uses
a polyfill if the `Object.assign` function is not available. I
pulled the polyfill implementation from:

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/assign

This is part of an effort to make Romo no longer dependent on jquery.