add a base js `parents` method
This is prep for moving off of jQuery and onto Romo's vanilla js
API.  This `parents` method emulatest jQuery's `parents` method,
getting a list of parent elems for a given child elem.  It also
optionally takes a selector string to filter the parents with.

Note: the diff is a little rough.  I moved the `closest` method
definition to jux it after the `parents` method since they both
look at parent elems and both take a selector.  This also jux's
the `prev|next` meths with the `select{Prev|Next}` meths.