(hotfix) turn on Romo custom events for event handling
Previously, we were using jQuery event handling under-the-covers
for backwards compatibility.  Since we are moving off of jQuery,
we need to stop doing this so we can make sure everything works
as expected off of jQuery events.

This is part of moving away from jQuery and on to vanilla js.