Fix binding and unbinding to events on the window/document
This fixes binding and unbinding to events on the window/document.
This is part of an effort to no longer use jquery. This updates
the `_el` function to detect if the elem is the `window` or
`document` and if so it doesn't try to read or write an attribute
on them because they don't support the attribute methods. Instead
it uses a hard-coded identifier (`"window"` or `"document"`) for
them. This allows them to work with the event binding and ensures
we can bind and unbind events as needed.