Update `RomoTooltip` component to not use jquery
This updates the `RomoTooltip` component to not use jquery. This
is part of the effort to no longer require jquery to use Romo.
This removes all of the jquery usage within the `RomoTooltip`
component. None of the other components initialize or use the
romo tooltip component.

This also updates the event names to be prefixed with
`romoTooltip` instead of just `tooltip` and renames variables as
well. This is part of having everything properly namespaced in
romo and ensuring it should work without issue with other
javascript libraries.