Update `RomoIndicatorTextInput` component to not use jquery
This updates the `RomoIndicatorTextInput` component to not use
jquery. This is part of the effort to no longer require jquery to
use Romo. This removes all of the jquery usage within the
`RomoIndicatorTextInput` component and also updates the currency
text input, datepicker, and option list dropdown components to
not use the jquery method for initializing a romo indicator text
input component.

This also updates the event names to be prefixed with
`romoIndicatorTextInput` instead of just `indicatorTextInput` and
renames variables as well. This is part of having everything
properly namespaced in romo and ensuring it should work without
issue with other javascript libraries.