Update `RomoInline` component to not use jquery
This updates the `RomoInline` component to not use jquery. This
is part of the effort to no longer require jquery to use Romo.
This removes all of the jquery usage within the `RomoInline`
component and also updates the inline form to not use the jquery
method for initializing a romo inline component.

This removes using jquery's `unbind` to remove all event handlers
from elems. The `Romo.off` doesn't support unbinding all event
handlers but we also decided that Romo doesn't need to be this
aggressive with its components elems. This allows users to bind
event handlers to the same events that romo inline does (though
it's not recommended).

This also updates the event names to be prefixed with `romoInline`
instead of just `inline` and renames variables as well. This is
part of having everything properly namespaced in romo and ensuring
it should work without issue with other javascript libraries.