Update `RomoOnkey` component to not use jquery
This updates the `RomoOnkey` component to not use jquery. This
is part of the effort to no longer require jquery to use Romo.
This removes all of the jquery usage within the `RomoOnkey`.

This also updates the event names to be prefixed with `romoOnkey`
instead of just `onkey` and renames variables as well. This is
part of having everything properly namespaced in romo and ensuring
it should work without issue with other javascript libraries.