Update `RomoSortable` component to not use jquery
This updates the `RomoSortable` component to not use jquery. This
is part of the effort to no longer require jquery to use Romo.
This removes all of the jquery usage within the `RomoSortable`
component. None of the other components initialize or use the
romo sortable component.

This also updates the event names to be prefixed with
`romoSortable` instead of just `sortable`. This is part of having
everything properly namespaced in romo and ensuring it should work
without issue with other javascript libraries.