Update `RomoOptionListDropdown` component to not use jquery
This updates the `RomoOptionListDropdown` component to not use
jquery. This is part of the effort to no longer require jquery to
use Romo. This removes all of the jquery usage within the
`RomoOptionListDropdown` component and also updates the picker and
select dropdown components to not use the jquery method for
initializing a romo option list dropdown component.

Note: This didn't update any event names because they are already
prefixed with `romoOptionListDropdown`.