Update `RomoSelectDropdown` component to not use jquery
This updates the `RomoSelectDropdown` component to not use jquery.
This is part of the effort to no longer require jquery to use
Romo. This removes all of the jquery usage within the
`RomoSelectDropdown` component and also updates the select
component to not use the jquery method for initializing a romo
select dropdown component.

This also updates the event names to be prefixed with
`romoSelectDropdown` instead of just `selectDropdown` and renames
variables as well. This is part of having everything properly
namespaced in romo and ensuring it should work without issue with
other javascript libraries.