rework the indicator component as a RomoSpinner component
The main reason for doing this is to avoid name conflicts with the
indicator text input.  This is also a better name b/c the
underlying class is also named "spinner" and this component wraps
that.

This not only renames the component and events, but also moves to
namespacing the events with "romo" per our latest convention.
This takes care of updating all of the other components that use
the spinner to rename their trigger events and usage to match.

Finally, this updates the RomoSpinner to not use jQuery logic. We
are moving away from depending on jQuery.  All logic has been
replaced with the Romo js equivalents.  I also reorganized the
methods to make all "private" methods have the leading underscore
per our convention.