Update `Romo.array` helper to cast non array-like values to arrays
This updates the `Romo.array` helper to cast non array-like values
to arrays. This is to avoid manually checking for arrays and
manually converting them to arrays if not.

The `Romo.array` helper now checks the value its passed and see
if it has a `length` property. If it has a length and can be
accessed like an array (via `[]`) then the object is array like
and is converted to an array. The helper also includes some
checks to avoid functions, strings, and the `window` object being
seen as an array like object because otherwise the helper would
try to convert them to an array. If the object is not array like
it is added as the single item in an array. Finally, if the value
is a `NodeList` or `HTMLCollection` then the helper includes a
short circuit to immediately convert them to an array. This
ensures the `Romo.f`, `Romo.find`, and `Romo.elems` helpers all
remain as fast as possible.

This also updates the select, picker, and the `Romo.triggerInitUI`
logic to no longer manually check for arrays and instead use the
`Romo.array` helper. This allows them to properly cast values to
arrays as needed.