Update `RomoModalForm` component to not use jquery
This updates the `RomoModalForm` component to not use jquery. This
is part of the effort to no longer require jquery to use Romo.
This removes all of the jquery usage within the `RomoModalForm`
component. None of the other components initialize or use the romo
modal form component.

This also updates the event names to be prefixed with
`romoModalForm` instead of just `modalForm`. This is part of
having everything properly namespaced in romo and ensuring it
should work without issue with other javascript libraries.

This also updates the `RomoForm` initialization, event names, and
variable names to match the updates to the `RomoForm` component.