base js: make event handling error if trying to use with elem collections
This is a protection feature.  Since Romo's js api is so similar
to jQuery's, it is logical that a dev might accidentally think to
call `.on|.off|.trigger` passing a collection of elems and have it
work.  This won't work.  So this add's a loud error to let you
know you are not using Romo's api correctly.  I

This should also protect us from mistakes as we port Romo from
jQuery to vanilla js.