Update `RomoInlineForm` component to not use jquery
This updates the `RomoInlineForm` component to not use jquery.
This is part of the effort to no longer require jquery to use
Romo. This removes all of the jquery usage within the
`RomoInlineForm` component. None of the other components
initialize or use the romo inline form component.

This also updates the event names to be prefixed with
`romoInlineForm` instead of just `inlineForm`. This is part of
having everything properly namespaced in romo and ensuring it
should work without issue with other javascript libraries.

This also updates the `RomoForm` initialization, event names, and
variable names to match the updates to the `RomoForm` component.