base js: remove `getComputedStyle` method
Now that we have the `.css` method (which is the exact same
implementation), we no longer need this method.  `css` is more
succinct and is what we'll be using going forward.

This removes the base js method and its usage in base js and
indicator text input.  The other uses have already been updated
as part of earlier cleanups/reworks in the effort to move off of
using jQuery.