Update append/prepend, before/after, and update to allow single elem
This updates the Romo helpers for updating or inserting elems to
allow passing a single elem. This is done by having them call
`Romo.array` on whatever they are passed and then using their
previous implementation which expected a set of elems. This was
noticed because dropdowns try to append their popup elem to a
popup parent and were erroring because it wasn't an array.

This also updates the `Romo.array` helper to short circuit if it's
passed an array object. This keeps the append/prepend,
before/after, and update fast when passed arrays of elems. This
probably should have been done in PR 177 when we originally
updated `Romo.array` to convert any value to an array and when
we added the short-circuit for `NodeList` and `HTMLCollection`.