Bug fixes and updates to latest conventions for sortable component
This updates the sortable component to fix some bugs introduced
while updating it to no longer use jquery. This also includes
updates for the sortable component to make it match some of our
latest conventions. This is part of updating Romo to no longer
require jquery.

This fixes the `Romo.before` call when the dragged elem is
dropped. It was accidentally updated to insert the placeholder
elem before the dragged elem instead of inserting the dragged
elem before the placeholder elem. This was incorrectly updated
in PR 161 when jquery was removed. It previously used
`insertBefore` on the dragged elem and is passed the placeholder
elem. This is the opposite order for how the elems needed to be
passed to `Romo.before` which is what caused the issue.

This updates the dragging, drag over, and placeholder class to
be lazy methods and also removes defaulting their values. This
is because the `addClass` and `removeClass` helpers error when
using an empty string which was the previous default. Now the
sortable component checks if it has the class before trying to
add or remove it which is consistent with these being optional.
This makes the methods lazy to match our latest conventions to
avoid all data attrs having to be set when the component is
initialized.

This updates the sortable to use `undefined` instead of `null`
for its attributes that are only set when dragging an element. The
other romo components have centered around using `undefined` so
the sortable component should as well. This also updates a few of
the `if` to check for `undefined` instead of using falsey values.
This is more explicit and should help avoid logic seemingly
working with invalid values.

This also fixes a few typos on method names. These were
accidentally introduced when the methods were renamed to our
private nameing scheme.