Update `Romo.children` and `Romo.siblings` to take optional selector
This updates the `Romo.children` and `Romo.siblings` helpers to
take an optional selector. This makes these helpers consistent
with the other helpers that find elements (like `parents` and
`closest`). This also fixes a bug with the select dropdown
component already expecting the `children` helper to accept and
use a selector. This was keeping selects from filtering their
options when their filter text input was changed.

This updates the `children` helper to take a selector and if
provided, it filters the child elems by using `Romo.is` to see
if they match the selector. This updates the `siblings` helper
to simply call the `children` helper and pass a selector. Now
that it's calling `children` it knows the result will be an
array so it can directly call `filter` on the result.