Update `Romo.offset`, don't use `document.body`
This updates `Romo.offset` to not use `document.body` to get the
scroll top or scroll left. In the latest versions of browsers
this has been deprecated and always returns 0. This fixes errors
with positioning dropdowns, tooltips, modals, and other
components when the viewport is scrolled.

This switches to using `window.pageYOffset` and
`window.pageXOffset`. This would have used `scrollY` and
`scrollX` which seems to be the preferred method for getting this
data but these are not supported in IE 9. The offset methods
are aliases for the scroll methods in non IE 9 browsers so this
should work in all browsers. Here's the MDN articles for
`scrollY` and `pageYOffset`:

https://developer.mozilla.org/en-US/docs/Web/API/Window/scrollY
https://developer.mozilla.org/en-US/docs/Web/API/Window/pageYOffset