Update dropdowns, modals, and tooltips to use `setTimeout` to place popup
This updates dropdowns, modals, and tooltips to use `setTimeout`
to place their popups. This fixes issues with the components
calculating their height/width appropriately when their content
is updated with components like pickers and selects. Both of these
change their HTML which causes their height to change. This in
turn causes the outer components height to change as well so
previous calculations may no longer be correct. By moving the
components into a `setTimeout` this allows the other components
time to adjust their HTML before placing the popup elem.

Note: This wasn't done on the popup open or resize event handler
for the components because it caused the dropdowns, modals, and
tooltips to have minor display issues. When the popup was opened
sometimes there would be a quick flash of the popup positioned
incorrectly. When resizing the animation wasn't smooth and it
would jitter as it adjusted its position based on the new window
size. This limits using the `setTimeout` to when the content of
the popup might change thus causing its height or width to
change.